# all: various internal optimizations