- apmhttp: move some HTTP context functions to internal/apmhttputil
- apmhttp: change responseWriter to not cache the CloseNotify or Flush methods, as that allocates more memory.
- apmhttp: use strings.Builder (go1.10+) or copy instead of fmt.Sprintf when formatting HTTP transaction names
- apmhttputil: pre-canonicalize HTTP headers (X-Real-IP -> X-Real-Ip) to avoid repeated string copies
- {apmhttputil,model}: format the full URL in the MarshalFastJSON method, to avoid allocating a string only to be copied into the buffer later.
- {elasticapm,model}: store time.Time in model objects, format using AppendFormat
- {elasticapm,model}: make Transaction.SpanCount a non-pointer field using fastjson's omitempty "isZero" method support

There's more to come, but they are more invasive, and change how we construct transaction context.

benchcmp for contrib/apmhttp:
```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkWithoutMiddleware//hello/world-4     608           560           -7.89%
BenchmarkWithoutMiddleware//sleep/1ms-4       1219645       1212475       -0.59%
BenchmarkWithMiddleware//hello/world-4        4471          3189          -28.67%
BenchmarkWithMiddleware//sleep/1ms-4          1197057       1214006       +1.42%

benchmark                                     old allocs     new allocs     delta
BenchmarkWithoutMiddleware//hello/world-4     3              3              +0.00%
BenchmarkWithoutMiddleware//sleep/1ms-4       0              0              +0.00%
BenchmarkWithMiddleware//hello/world-4        24             15             -37.50%
BenchmarkWithMiddleware//sleep/1ms-4          20             11             -45.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkWithoutMiddleware//hello/world-4     74            74            +0.00%
BenchmarkWithoutMiddleware//sleep/1ms-4       0             0             +0.00%
BenchmarkWithMiddleware//hello/world-4        1308          976           -25.38%
BenchmarkWithMiddleware//sleep/1ms-4          1204          865           -28.16%
```